Tree Chopper changes + Leaf Breaker enchantment
Tree Chopper scan is now fuzzy.. mean it'll break any block that matched within it's scan area.
This allows it to break the entire giant mushroom instead of just the stem. This also means that if trees are planted in a big group, it can take down all of the trees regardless of wood type.
Tree Chopper doesn't require leaves to be attached to the stem anymore.

Leaf Breaker Enchantment added to allow the breaking of leaves.